Refactor oodns to become dnsclient/oodnsclient
So, the current idea is to use the same pattern used by the
dnstransports, and have different clients one can compose.

The most obvious, initial configuration is the one where one
is decomposing doing the real business and emitting events.
