+++ internal/dnsclient/oodnsclient/oodnsclient.go
@@ -1,5 +1,5 @@
-// Package oodns is OONI's DNS client.
-package oodns
+// Package oodnsclient is OONI's DNS client.
+package oodnsclient
 
 import (
 	"context"
@@ -21,8 +21,8 @@ type Client struct {
 	transport dnsx.RoundTripper
 }
 
-// NewClient creates a new OONI DNS client instance.
-func NewClient(
+// New creates a new OONI DNS client instance.
+func New(
 	beginning time.Time, handler model.Handler, t dnsx.RoundTripper,
 ) *Client {
 	return &Client{
@@ -78,12 +78,10 @@ func (c *Client) LookupHost(ctx context.Context, hostname string) ([]string, err
 			}
 		}
 	}
-	return LookupHostResult(addrs, errA, errAAAA)
+	return lookupHostResult(addrs, errA, errAAAA)
 }
 
-// LookupHostResult computes the final result of LookupHost. You generally
-// only care about this function when writing tests.
-func LookupHostResult(addrs []string, errA, errAAAA error) ([]string, error) {
+func lookupHostResult(addrs []string, errA, errAAAA error) ([]string, error) {
 	if len(addrs) > 0 {
 		return addrs, nil
 	}
@@ -118,7 +116,7 @@ func (c *Client) newQueryWithQuestion(q dns.Question) (query *dns.Msg) {
 }
 
 func (c *Client) roundTrip(ctx context.Context, query *dns.Msg) (reply *dns.Msg, err error) {
-	return c.RoundTripEx(
+	return c.doRoundTrip(
 		ctx, query, func(msg *dns.Msg) ([]byte, error) {
 			return msg.Pack()
 		},
@@ -131,9 +129,7 @@ func (c *Client) roundTrip(ctx context.Context, query *dns.Msg) (reply *dns.Msg,
 	)
 }
 
-// RoundTripEx is a mockable implementation of the piece
-// of code that performs the DNS round trip.
-func (c *Client) RoundTripEx(
+func (c *Client) doRoundTrip(
 	ctx context.Context,
 	query *dns.Msg,
 	pack func(msg *dns.Msg) ([]byte, error),
