+++ internal/dnsclient/oodnsclient/oodnsclient_test.go
@@ -1,4 +1,4 @@
-package oodns_test
+package oodnsclient
 
 import (
 	"context"
@@ -12,7 +12,6 @@ import (
 	"github.com/ooni/netx/dnsx"
 	"github.com/ooni/netx/handlers"
 	"github.com/ooni/netx/internal/dnstransport/dnsovertcp"
-	"github.com/ooni/netx/internal/oodns"
 )
 
 func newtransport() dnsx.RoundTripper {
@@ -25,7 +24,7 @@ func newtransport() dnsx.RoundTripper {
 }
 
 func TestLookupAddr(t *testing.T) {
-	client := oodns.NewClient(
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
 	addrs, err := client.LookupAddr(context.Background(), "130.192.91.211")
@@ -38,7 +37,7 @@ func TestLookupAddr(t *testing.T) {
 }
 
 func TestLookupCNAME(t *testing.T) {
-	client := oodns.NewClient(
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
 	addrs, err := client.LookupCNAME(context.Background(), "www.ooni.io")
@@ -51,7 +50,7 @@ func TestLookupCNAME(t *testing.T) {
 }
 
 func TestLookupHost(t *testing.T) {
-	client := oodns.NewClient(
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
 	addrs, err := client.LookupHost(context.Background(), "www.google.com")
@@ -64,7 +63,7 @@ func TestLookupHost(t *testing.T) {
 }
 
 func TestLookupNonexistent(t *testing.T) {
-	client := oodns.NewClient(
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
 	addrs, err := client.LookupHost(context.Background(), "nonexistent.ooni.io")
@@ -77,7 +76,7 @@ func TestLookupNonexistent(t *testing.T) {
 }
 
 func TestLookupMX(t *testing.T) {
-	client := oodns.NewClient(
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
 	addrs, err := client.LookupMX(context.Background(), "ooni.io")
@@ -90,7 +89,7 @@ func TestLookupMX(t *testing.T) {
 }
 
 func TestLookupNS(t *testing.T) {
-	client := oodns.NewClient(
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
 	addrs, err := client.LookupNS(context.Background(), "ooni.io")
@@ -102,11 +101,11 @@ func TestLookupNS(t *testing.T) {
 	}
 }
 
-func TestRoundTripExPackFailure(t *testing.T) {
-	client := oodns.NewClient(
+func TestDoRoundTripPackFailure(t *testing.T) {
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
-	_, err := client.RoundTripEx(
+	_, err := client.doRoundTrip(
 		context.Background(), nil,
 		func(msg *dns.Msg) ([]byte, error) {
 			return nil, errors.New("mocked error")
@@ -123,11 +122,11 @@ func TestRoundTripExPackFailure(t *testing.T) {
 	}
 }
 
-func TestRoundTripExRoundTripFailure(t *testing.T) {
-	client := oodns.NewClient(
+func TestDoRoundTripRoundTripFailure(t *testing.T) {
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
-	_, err := client.RoundTripEx(
+	_, err := client.doRoundTrip(
 		context.Background(), nil,
 		func(msg *dns.Msg) ([]byte, error) {
 			return nil, nil
@@ -144,11 +143,11 @@ func TestRoundTripExRoundTripFailure(t *testing.T) {
 	}
 }
 
-func TestRoundTripExUnpackFailure(t *testing.T) {
-	client := oodns.NewClient(
+func TestDoRoundTripUnpackFailure(t *testing.T) {
+	client := New(
 		time.Now(), handlers.NoHandler, newtransport(),
 	)
-	_, err := client.RoundTripEx(
+	_, err := client.doRoundTrip(
 		context.Background(), nil,
 		func(msg *dns.Msg) ([]byte, error) {
 			return nil, nil
@@ -166,7 +165,7 @@ func TestRoundTripExUnpackFailure(t *testing.T) {
 }
 
 func TestLookupHostResultNoName(t *testing.T) {
-	addrs, err := oodns.LookupHostResult(nil, nil, nil)
+	addrs, err := lookupHostResult(nil, nil, nil)
 	if err == nil {
 		t.Fatal("expected an error here")
 	}
@@ -176,7 +175,7 @@ func TestLookupHostResultNoName(t *testing.T) {
 }
 
 func TestLookupHostResultAAAAError(t *testing.T) {
-	addrs, err := oodns.LookupHostResult(nil, nil, errors.New("mocked error"))
+	addrs, err := lookupHostResult(nil, nil, errors.New("mocked error"))
 	if err == nil {
 		t.Fatal("expected an error here")
 	}
