+++ internal/dnsconf/dnsconf.go
@@ -14,7 +14,7 @@ import (
 	"github.com/ooni/netx/internal/dnstransport/dnsovertcp"
 	"github.com/ooni/netx/internal/dnstransport/dnsoverudp"
 	"github.com/ooni/netx/internal/httptransport"
-	"github.com/ooni/netx/internal/oodns"
+	"github.com/ooni/netx/internal/dnsclient/oodnsclient"
 	"github.com/ooni/netx/model"
 )
 
@@ -93,5 +93,5 @@ func NewResolver(
 	if transport == nil {
 		return nil, errors.New("dnsconf: unsupported network value")
 	}
-	return oodns.NewClient(dialer.Beginning, dialer.Handler, transport), nil
+	return oodnsclient.New(dialer.Beginning, dialer.Handler, transport), nil
 }
